requirements.txt
gradio
qrcode
pillow
pypdf
python-barcode
